add search classes functionality to client page
